+++ src/js/config.js
@@ -2,4 +2,8 @@
 
 // so if want to change some configurations we do it using the config
 
+import 'babel-polyfill';
+
 export const API_URL = 'https://forkify-api.herokuapp.com/api/v2/recipes';
+
+export const TIMEOUT_SEC = 10;
