+++ src/js/controller.js
@@ -8,14 +8,7 @@ import icons from '../img/icons.svg';
 
 // const recipeContainer = document.querySelector('.recipe');
 
-// This is promisifyin a setTimeout callback function
-const timeout = function (s) {
-  return new Promise(function (_, reject) {
-    setTimeout(function () {
-      reject(new Error(`Request took too long! Timeout after ${s} second`));
-    }, s * 1000);
-  });
-};
+
 
 // https://forkify-api.herokuapp.com/v2
